Light 3종 추가

+++ Engine/TableDescriptorHeap.h
@@ -20,7 +20,7 @@ public:
 	D3D12_CPU_DESCRIPTOR_HANDLE GetCPUHandle(SRV_REGISTER reg);
 
 private:
-	D3D12_CPU_DESCRIPTOR_HANDLE GetCPUHandle(uint32 reg);
+	D3D12_CPU_DESCRIPTOR_HANDLE GetCPUHandle(uint8 reg);
 
 private:
